Refactor DOcDB commands into separate file
